[feat/#42] 문화생활 기록 단일 조회 API 연동

+++ app/src/main/java/com/example/baba/ui/home/CreateActivity.kt
@@ -63,7 +63,7 @@ class CreateActivity : ComponentActivity() {
                         category = initialCategory,
                         date = initialDate,
                         title = title,
-                        rating = rating.toFloat(),
+                        rating = rating,
                         review = review,
                         onReviewChange = { review = it },
                         location = location,
